pyproject.toml
[project]
name = "object_detection_comparison"
version = "0.1.0"
description = "Comparative analysis of different object detection models"
authors = [{ name = "YgLK", email = "jkbszpunar@gmail.com" }]
requires-python = ">=3.10,<3.12"
readme = "README.md"

dependencies = [
    "hydra-core>=1.3.2,<2",
    "hydra-colorlog>=1.2.0,<2",
    "pydantic>=2.0.0,<3",
    "numpy>=1.24.0,<2",
    "matplotlib>=3.7.0,<4",
    "seaborn>=0.12.0,<0.13",
    "tqdm>=4.65.0,<5",
    "opencv-python-headless>=4.7.0,<5",
    "albumentations>=1.3.0,<2",
    "pycocotools>=2.0.6,<3",
    "tensorboard>=2.12.0,<3",
    "mlflow>=2.3.0,<3",
    "omegaconf>=2.3.0,<3",
    "jupyter>=1.0.0,<2",
    "ultralytics>=8.3.94,<9",
    "loguru>=0.7.3",
    "dill>=0.4.0",
    "jinja2>=3.1.6",
    "pyyaml>=6.0.2",
    "pandas>=2.2.3",
    "pyarrow>=19.0.1",
    "pillow>=8.3.0,<12",
    "click>=8.1.8",
    "openmim>=0.3.9",
    "torch>=2.6.0",
    "wandb>=0.21.0",
    "python-dotenv>=1.1.1",
    "pre-commit>=4.2.0",
    "nbstripout>=0.8.1",
    "pytorch-lightning>=2.5.2",
    "torchmetrics>=1.7.4",
    "psutil>=7.0.0",
    "rfdetr[metrics]>=1.1.0",
    "optuna-dashboard>=0.19.0",
    "tidecv>=1.0.1",
    "ensemble-boxes>=1.0.9",
    "pysqlite3-binary>=0.5.4",
    "jupytext>=1.17.3",
]

[dependency-groups]
dev = [
    "pytest>=7.3.1,<8",
    "pytest-cov>=4.1.0,<5",
    "pytest-mock>=3.10.0,<4",
    "coverage>=7.2.7,<8",
    "ruff>=0.0.262,<0.0.263",
    "mypy>=1.2.0,<2",
    "types-PyYAML>=6.0.12,<7",
    "types-Pillow>=10.0.0,<11",
]

[tool.hatch.build.targets.sdist]
include = ["odc"]

[tool.hatch.build.targets.wheel]
include = ["odc"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = [
    "A",    # prevent using keywords that clobber python builtins
    "I",    # isort - import sorting,
    "T",    # typing
    "ISC",  # implicit string concatenation
    "UP",   # alert you when better syntax is available in your python version
]
ignore = []

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F403", "F401", "I001"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
known-first-party = ["odc"]
force-sort-within-sections = true
lines-after-imports = 2

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "--strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
